+++ app/src/main/java/com/m6code/driza/model/Datum.java
@@ -3,11 +3,13 @@ package com.m6code.driza.model;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigInteger;
+
 public class Datum {
 
     @SerializedName("id")
     @Expose
-    private Integer id;
+    private BigInteger id;
     @SerializedName("readable")
     @Expose
     private Boolean readable;
@@ -103,7 +105,7 @@ public class Datum {
     private Integer nbAlbum;
     @SerializedName("nb_fan")
     @Expose
-    private Integer nbFan;
+    private BigInteger nbFan;
     @SerializedName("radio")
     @Expose
     private Boolean radio;
@@ -133,11 +135,11 @@ public class Datum {
 
 
 
-    public Integer getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
 
@@ -373,11 +375,11 @@ public class Datum {
         this.nbAlbum = nbAlbum;
     }
 
-    public Integer getNbFan() {
+    public BigInteger getNbFan() {
         return nbFan;
     }
 
-    public void setNbFan(Integer nbFan) {
+    public void setNbFan(BigInteger nbFan) {
         this.nbFan = nbFan;
     }
